Class files
The Module and Shell classes are now individual files.  Additionally, the CalculateVelocity() method now functions correctly.  The issue was that I was mixing up the indices of the module list with the actual values at those indices (which indicate the quantity of the module at that index).  The KD and AP functions still need to be corrected and completed.