Added DPS per Volume
The goal is to increase the usefulness of the numbers by making them represent practical units.  A gun is not made exclusively of intakes - it also has cooling, recoil absorption, and railgun chargers, all of which scale linearly with the number of intakes.  There are other components as well, but these are constant - gauge increasers, firing piece, rail magnets - and so do not scale with number of intakes.
DPS amounts now incorporate the volume of the intake and the volume of a loader or clip attached to the intake.  I plan to add cooling, recoil absorption, and charger volume next.